Revert "Enhancement tokensearcher"
