+++ .github/workflows/tests.yml
@@ -19,4 +19,4 @@ jobs:
         architecture: x64
     - run: pip install nox==2020.8.22
     - run: pip install poetry==1.1.4
-    - run: nox
+    - run: nox -s lint mypy safety tests typeguard -p ${{ matrix.python-verison }}
