+++ docs/reference.rst
@@ -6,6 +6,27 @@ Reference
     :backlinks: none
 
 
+spaczz.fuzz
+--------------------------
+
+.. automodule:: spaczz.fuzz
+   :members:
+
+
+spaczz.regex
+--------------------------
+
+.. automodule:: spaczz.regex
+   :members:
+
+
+spaczz.similarity
+--------------------------
+
+.. automodule:: spaczz.similarity
+   :members:
+
+
 spaczz.matcher
 --------------------------
 
@@ -20,17 +41,10 @@ spaczz.pipeline
    :members:
 
 
-spaczz.search
+spaczz.process
 --------------------------
 
-.. automodule:: spaczz.search
-   :members:
-
-
-spaczz.regex
---------------------------
-
-.. automodule:: spaczz.regex
+.. automodule:: spaczz.process
    :members:
 
 
@@ -38,11 +52,4 @@ spaczz.util
 --------------------------
 
 .. automodule:: spaczz.util
-   :members:
-
-
-spaczz.exceptions
---------------------------
-
-.. automodule:: spaczz.exceptions
    :members:
