+++ src/spaczz/search/__init__.py
@@ -3,12 +3,5 @@ from ._phrasesearcher import _PhraseSearcher
 from .fuzzysearcher import FuzzySearcher
 from .regexsearcher import RegexSearcher
 from .similaritysearcher import SimilaritySearcher
-from .tokensearcher import TokenSearcher
 
-__all__ = [
-    "_PhraseSearcher",
-    "FuzzySearcher",
-    "RegexSearcher",
-    "SimilaritySearcher",
-    "TokenSearcher",
-]
+__all__ = ["_PhraseSearcher", "FuzzySearcher", "RegexSearcher", "SimilaritySearcher"]
